bookingandmanagementapis/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
colorama==0.4.6
cryptography==44.0.2
Django==5.1.7
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
